Move schema & validate API data
