+++ src/schema/author.ts
@@ -1,0 +1,13 @@
+import { z } from "astro:content";
+
+export const authorSchema = z
+  .object({
+    name: z.string(),
+    avatar: z.string().url().optional(),
+    twitter: z.string().url().optional(),
+    github: z.string().url().optional(),
+    website: z.string().url().optional(),
+  })
+  .strict();
+
+export type AuthorFrontmatter = z.infer<typeof authorSchema>;
