--- src/utils/types.ts
@@ -1,6 +1,0 @@
-import type { BlogFrontmatter } from "@content/_schemas";
-
-export type PostSearch = {
-  href: string;
-  frontmatter: BlogFrontmatter;
-};
