+++ tsconfig.json
@@ -10,7 +10,8 @@
       "@content/*": ["content/*"],
       "@layouts/*": ["layouts/*"],
       "@pages/*": ["pages/*"],
-      "@utils/*": ["utils/*"]
+      "@utils/*": ["utils/*"],
+      "@schema/*": ["schema/*"]
     }
   }
 }
